converted unit test into plugin
